+++ src/Shortcut/Alipay/AuthShortcut.php
@@ -1,0 +1,74 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yansongda\Pay\Shortcut\Alipay;
+
+use Yansongda\Artful\Contract\ShortcutInterface;
+use Yansongda\Artful\Exception\InvalidParamsException;
+use Yansongda\Artful\Plugin\ParserPlugin;
+use Yansongda\Pay\Exception\Exception;
+use Yansongda\Pay\Plugin\Alipay\V2\AddPayloadSignaturePlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\AddRadarPlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\FormatPayloadBizContentPlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\Open\Authorization\TokenAppPlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\Open\Authorization\TokenAppQueryPlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\ResponsePlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\StartPlugin;
+use Yansongda\Pay\Plugin\Alipay\V2\VerifySignaturePlugin;
+use Yansongda\Supports\Str;
+
+class AuthShortcut implements ShortcutInterface
+{
+    /**
+     * @param array<string, mixed> $params
+     *
+     * @return array<class-string>
+     *
+     * @throws InvalidParamsException
+     */
+    public function getPlugins(array $params): array
+    {
+        $method = Str::camel($params['_action'] ?? 'token_app').'Plugins';
+
+        if (method_exists($this, $method)) {
+            return $this->{$method}();
+        }
+
+        throw new InvalidParamsException(Exception::PARAMS_SHORTCUT_ACTION_INVALID, "您所提供的 action 方法 [{$method}] 不支持，请参考文档或源码确认");
+    }
+
+    /**
+     * @return array<class-string>
+     */
+    protected function tokenAppPlugins(): array
+    {
+        return [
+            StartPlugin::class,
+            TokenAppPlugin::class,
+            FormatPayloadBizContentPlugin::class,
+            AddPayloadSignaturePlugin::class,
+            AddRadarPlugin::class,
+            VerifySignaturePlugin::class,
+            ResponsePlugin::class,
+            ParserPlugin::class,
+        ];
+    }
+
+    /**
+     * @return array<class-string>
+     */
+    protected function queryPlugins(): array
+    {
+        return [
+            StartPlugin::class,
+            TokenAppQueryPlugin::class,
+            FormatPayloadBizContentPlugin::class,
+            AddPayloadSignaturePlugin::class,
+            AddRadarPlugin::class,
+            VerifySignaturePlugin::class,
+            ResponsePlugin::class,
+            ParserPlugin::class,
+        ];
+    }
+}